+++ custom_components/unifi_insights/config_flow.py
@@ -5,7 +5,6 @@ from __future__ import annotations
 import logging
 from typing import Any
 
-import voluptuous as vol
 from homeassistant.config_entries import (
     ConfigEntry,
     ConfigFlow,
@@ -21,6 +20,7 @@ from homeassistant.helpers.selector import (
     SelectSelectorMode,
 )
 from pydantic import ValidationError
+import voluptuous as vol
 
 from .api import (
     ApiKeyAuth,